Add heuristic test
