--- play-ahc-ws-standalone/src/test/scala/play/api/libs/ws/ahc/cache/CacheAsyncHttpProviderSpec.scala
@@ -1,20 +1,0 @@
-/*
- * Copyright (C) 2009-2017 Lightbend Inc. <https://www.lightbend.com>
- *
- */
-
-package play.api.libs.ws.ahc.cache
-
-import org.specs2.mutable.Specification
-
-class CacheAsyncHttpProviderSpec extends Specification with CacheBuilderMethods {
-
-  "CacheAsyncHttpProvider.execute" should {
-
-    "work" in {
-      pending
-    }
-
-  }
-
-}
