--- play-ahc-ws-standalone/src/test/scala/play/api/libs/ws/ahc/cache/CacheBuilderMethods.scala
@@ -1,51 +1,0 @@
-/*
- * Copyright (C) 2009-2017 Lightbend Inc. <https://www.lightbend.com>
- *
- */
-
-package play.api.libs.ws.ahc.cache
-
-import akka.stream.Materializer
-import play.shaded.ahc.io.netty.handler.codec.http.{ DefaultHttpHeaders, HttpHeaders }
-import play.shaded.ahc.org.asynchttpclient.{ Request, RequestBuilder }
-
-import scala.collection.mutable
-import scala.concurrent.ExecutionContext
-
-/**
- * Utility methods to make building requests and responses easier.
- */
-trait CacheBuilderMethods {
-
-  def generateCache: AhcHttpCache = {
-    import ExecutionContext.Implicits.global
-    AhcHttpCache(new StubHttpCache(), false)
-  }
-
-  def generateRequest(url: String)(block: HttpHeaders => HttpHeaders): Request = {
-    val requestBuilder = new RequestBuilder()
-    val requestHeaders = block(new DefaultHttpHeaders())
-
-    requestBuilder
-      .setUrl(url)
-      .setHeaders(requestHeaders)
-      .build
-  }
-
-}
-
-class StubHttpCache extends Cache {
-
-  private val underlying = new mutable.HashMap[EffectiveURIKey, ResponseEntry]()
-
-  override def remove(key: EffectiveURIKey): Unit = underlying.remove(key)
-
-  override def put(key: EffectiveURIKey, entry: ResponseEntry): Unit = underlying.put(key, entry)
-
-  override def get(key: EffectiveURIKey): ResponseEntry = underlying.get(key).orNull
-
-  override def close(): Unit = {
-
-  }
-
-}
